auth_service/requirements.txt
fastapi
SQLAlchemy
pydantic
Alembic
asyncpg
psycopg2-binary
databases
pydantic_settings
uvicorn
passlib
passlib[bcrypt]
pyjwt
pydantic[email]
fastapi_mail
redis
loguru
python-jose
 python-multipart
